# GTFS-EXPRESS-API/docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: GTFS Express API
  version: 1.0.0
  description: |-
    Professional GTFS Schedule editor and validator — REST surface of the GTFS Express backend.

    ## Session model

    The API is **session-based**: `POST /gtfs/upload` (or `GET /gtfs/load-sample`) creates an
    isolated, in-memory-style SQLite database under `uploads/{sessionId}/gtfs.db` and returns a
    UUIDv4 `sessionId`. Every subsequent call targets that database through the `X-Session-ID`
    header. There is **no global persistent database**: sessions expire after a TTL (2 h dev / 1 h
    prod) and the server enforces a concurrent-session cap (HTTP 503 beyond it). Work is persisted
    by re-exporting the feed as a GTFS ZIP (`GET /gtfs/edit/export`) or saving a `.gtfsproj`
    snapshot (`GET /gtfs/edit/project/export`).

    ## Validation engine

    Validation is performed by a single engine: the **MobilityData Canonical Validator** (Java JAR,
    embedded in the Docker image). Reports are normalised to the internal shape documented in the
    `ValidationReport` schema; `POST /gtfs/edit/validate.canonical` returns the untouched
    canonical JSON instead. Severity profiles (canonical / strict / lenient / fr-datagouv) only
    override severities — the rule set itself stays aligned with MobilityData.

    ## Edit CRUD pattern

    All `/gtfs/edit/{table}` families (stops, routes, trips, stop_times, calendar, fares v1/v2,
...
    envelope, documented once and referenced everywhere:

...

...

...

...
servers:
...
paths:
...
  /gtfs/edit/routes/{route_id}:
    patch:
      tags:
        - Edit
      summary: Update a routes record.
      description: Generic edit CRUD endpoint — request/response envelope described in the **Edit CRUD
        pattern** section of the API description.
      parameters:
        - name: route_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        $ref: "#/components/requestBodies/EntityPatch"
      security:
        - SessionId: []
      responses:
        "200":
          $ref: "#/components/responses/EditOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/EditModeRequired"
    delete:
      tags:
        - Edit
      summary: Delete a routes record.
      description: Generic edit CRUD endpoint — request/response envelope described in the **Edit CRUD
        pattern** section of the API description.
      parameters:
        - name: route_id
          in: path
          required: true
          schema:
            type: string
      security:
        - SessionId: []
      responses:
        "200":
          $ref: "#/components/responses/EditOk"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/EditModeRequired"
...
components:
  securitySchemes:
    SessionId:
      type: apiKey
      in: header
      name: X-Session-ID
      description: UUIDv4 returned by POST /gtfs/upload or GET /gtfs/load-sample. Required by every
        session-scoped route; strictly validated server-side (path-traversal guard).
    BetaCode:
      type: apiKey
      in: header
      name: X-Beta-Code
      description: Beta access code. Required by edit-mode entry, project import and AI routes while the
        beta gate is enabled (no-op when BETA_GATE_DISABLED=true).
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Operator token (ADMIN_TOKEN env var) for the /gtfs/admin/* dashboard endpoints.
  responses:
...
    EditOk:
      description: Mutation applied and logged in the session edit history.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EditResult"
...
    BadRequest:
      description: "Invalid input: missing/invalid X-Session-ID, validation failure, or malformed body."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    EditModeRequired:
      description: Rejected by an access guard. Returned as 403 when the beta gate denies the X-Beta-Code
        or a mutation is attempted on the read-only SQL surface; edit handlers reject sessions that
        are not in edit mode with 409 and code SESSION_NOT_IN_EDIT_MODE.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Session, feed or record not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  requestBodies:
...
    EntityPatch:
      description: Partial JSON map of GTFS column names to values. Only whitelisted editable columns are
        applied; primary-key mutation is forbidden.
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
...
  schemas:
    Error:
      type: object
      description: Standard error envelope. `message`, `code` and `details` are present on a subset of
        endpoints.
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error (or machine code on gate endpoints).
        message:
          type: string
          description: Optional human-readable detail.
        code:
          type: string
          description: Optional machine-readable code (e.g. SESSION_NOT_IN_EDIT_MODE).
        details:
          type: array
          items: {}
          description: Optional structured validation violations.
    EditResult:
      type: object
      description: "Envelope returned by every edit mutation. The mutated row is echoed under a key named
        after the entity (e.g. `stop`, `route`, `trip`); updates add the `changed` column list,
        deletes return the primary key under `deleted`. The `validation` envelope is always present
        and currently always `skipped: true` — full canonical validation runs on demand via POST
        /gtfs/edit/validate."
      properties:
        changed:
          type: array
          items:
            type: string
          description: Columns actually modified (updates only).
        deleted:
          type: string
          description: Primary key of the removed record (deletes only).
        validation:
          type: object
          description: Per-edit incremental validation envelope (skipped under the canonical-only engine).
          properties:
            items:
              type: array
              items:
                type: object
                additionalProperties: true
            skipped:
              type: boolean
            reason:
              type: string
              example: incremental_validation_disabled_canonical_only
            elapsedMs:
              type: number
            truncated:
              type: integer
            totalAvailable:
              type: integer
      additionalProperties: true
...